0803-vue-blog-drf/vue_blog_api/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid==2.4.2
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
Django @ file:///tmp/build/80754af9/django_1605066529526/work
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httpie==2.3.0
httpie-jwt-auth==0.4.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
PySocks==1.7.1
pytz==2020.1
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
